% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_Functions.R,
%   R/oxy-RandomForest_Functions.R
\name{cm.aggregate}
\alias{cm.aggregate}
\title{FUNCTION_TITLE}
\usage{
cm.aggregate(cm)

cm.aggregate(cm)
}
\arguments{
\item{cm}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
