% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_Functions.R,
%   R/oxy-RandomForest_Functions.R
\name{gdens}
\alias{gdens}
\title{FUNCTION_TITLE}
\usage{
gdens(df, x, dvar, tsize = 24)

gdens(df, x, dvar, tsize = 24)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{x}{PARAM_DESCRIPTION}

\item{dvar}{PARAM_DESCRIPTION}

\item{tsize}{PARAM_DESCRIPTION, Default: 24}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
