% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_Functions.R,
%   R/oxy-RandomForest_Functions.R
\name{mean_var.rfmodel}
\alias{mean_var.rfmodel}
\title{FUNCTION_TITLE}
\usage{
mean_var.rfmodel(rf.type, performance.table)

mean_var.rfmodel(rf.type, performance.table)
}
\arguments{
\item{rf.type}{PARAM_DESCRIPTION}

\item{performance.table}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
