% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_Functions.R,
%   R/oxy-RandomForest_Functions.R
\name{mtry.guide}
\alias{mtry.guide}
\title{Recommended mtry range}
\usage{
mtry.guide(npred, rf.type, spread.upper = 5, spread.lower = 4)

mtry.guide(npred, rf.type, spread.upper = 5, spread.lower = 4)
}
\arguments{
\item{npred}{PARAM_DESCRIPTION}

\item{rf.type}{PARAM_DESCRIPTION}

\item{spread.upper}{PARAM_DESCRIPTION, Default: 5}

\item{spread.lower}{PARAM_DESCRIPTION, Default: 4}
}
\value{
Returns a list: \cr
 \code{ideal.mtry} is optimal mtry based on tree type and number of predictor variables \cr
 \code{upper.mtry} is upper range of mtry for downstream hyperparameter tuning \cr
 \code{lower.mtry} is lowest value of mtry for downstream hyperparameter tuning \cr
 \code{range} is range of mtry values to try in downstream hyperparameter tuning given in the format: \code{upper.mtry:lower.mtry}

OUTPUT_DESCRIPTION
}
\description{
Recommended mtry range based on number of paramerets and tree type

FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
