% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_Functions.R,
%   R/oxy-RandomForest_Functions.R
\name{multi.density.plot}
\alias{multi.density.plot}
\title{FUNCTION_TITLE}
\usage{
multi.density.plot(
  df,
  response,
  tsize = 24,
  top = 20,
  varimp,
  rf.type,
  metric,
  scale = 0.8,
  ncol = 2,
  xlab = NA,
  extract.names.df = NA
)

multi.density.plot(
  df,
  response,
  tsize = 24,
  top = 20,
  varimp,
  rf.type,
  metric,
  scale = 0.8,
  ncol = 2,
  xlab = NA,
  extract.names.df = NA
)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{response}{PARAM_DESCRIPTION}

\item{tsize}{PARAM_DESCRIPTION, Default: 24}

\item{top}{PARAM_DESCRIPTION, Default: 20}

\item{varimp}{PARAM_DESCRIPTION}

\item{rf.type}{PARAM_DESCRIPTION}

\item{metric}{PARAM_DESCRIPTION}

\item{scale}{PARAM_DESCRIPTION, Default: 0.8}

\item{ncol}{PARAM_DESCRIPTION, Default: 2}

\item{xlab}{PARAM_DESCRIPTION, Default: NA}

\item{extract.names.df}{PARAM_DESCRIPTION, Default: NA}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[cowplot]{plot_grid}}

\code{\link[cowplot]{plot_grid}}
}
