% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_Functions.R,
%   R/oxy-RandomForest_Functions.R
\name{sa_rf}
\alias{sa_rf}
\title{FUNCTION_TITLE}
\usage{
sa_rf(
  vpred,
  mtry = 1:10,
  ntree = (1:10) * 500,
  wd = wd.in,
  custom.name = FALSE,
  date = "",
  dataframe.name = "",
  predictors.name = "",
  nset = 10
)

sa_rf(
  vpred,
  mtry = 1:10,
  ntree = (1:10) * 500,
  wd = wd.in,
  custom.name = FALSE,
  date = "",
  dataframe.name = "",
  predictors.name = "",
  nset = 10
)
}
\arguments{
\item{vpred}{PARAM_DESCRIPTION}

\item{mtry}{PARAM_DESCRIPTION, Default: 1:10}

\item{ntree}{PARAM_DESCRIPTION, Default: (1:10) * 500}

\item{wd}{PARAM_DESCRIPTION, Default: wd.in}

\item{custom.name}{PARAM_DESCRIPTION, Default: FALSE}

\item{date}{PARAM_DESCRIPTION, Default: ''}

\item{dataframe.name}{PARAM_DESCRIPTION, Default: ''}

\item{predictors.name}{PARAM_DESCRIPTION, Default: ''}

\item{nset}{PARAM_DESCRIPTION, Default: 10}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[e1071]{tune.wrapper}}
 \code{\link[randomForest]{importance}}

\code{\link[e1071]{tune.wrapper}}
 \code{\link[randomForest]{importance}}
}
