% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_Functions.R,
%   R/oxy-RandomForest_Functions.R
\name{select.varimp}
\alias{select.varimp}
\title{FUNCTION_TITLE}
\usage{
\method{select}{varimp}(rf.type, metric)

\method{select}{varimp}(rf.type, metric)
}
\arguments{
\item{rf.type}{PARAM_DESCRIPTION}

\item{metric}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
