% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_Functions.R,
%   R/oxy-RandomForest_Functions.R
\name{split.ratio}
\alias{split.ratio}
\title{FUNCTION_TITLE}
\usage{
\method{split}{ratio}(df, nsets = 10, train.ratio = 0.8, rf.type, vpred)

\method{split}{ratio}(df, nsets = 10, train.ratio = 0.8, rf.type, vpred)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{nsets}{PARAM_DESCRIPTION, Default: 10}

\item{train.ratio}{PARAM_DESCRIPTION, Default: 0.8}

\item{rf.type}{PARAM_DESCRIPTION}

\item{vpred}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
