% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_Functions.R,
%   R/oxy-RandomForest_Functions.R
\name{varimport_plot}
\alias{varimport_plot}
\title{FUNCTION_TITLE}
\usage{
varimport_plot(
  varimp,
  rf.type,
  metric = "gini",
  selection_type = "random_top",
  top = 20,
  xlab = NA,
  extract.names.df = NA
)

varimport_plot(
  varimp,
  rf.type,
  metric = "gini",
  selection_type = "random_top",
  top = 20,
  xlab = NA,
  extract.names.df = NA
)
}
\arguments{
\item{varimp}{PARAM_DESCRIPTION}

\item{rf.type}{PARAM_DESCRIPTION}

\item{metric}{PARAM_DESCRIPTION, Default: 'gini'}

\item{selection_type}{PARAM_DESCRIPTION, Default: 'random_top'}

\item{top}{PARAM_DESCRIPTION, Default: 20}

\item{xlab}{PARAM_DESCRIPTION, Default: NA}

\item{extract.names.df}{PARAM_DESCRIPTION, Default: NA}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{count}}

\code{\link[dplyr]{count}}
}
