% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_Functions.R
\name{mtry.guide}
\alias{mtry.guide}
\title{Recommended mtry range}
\usage{
mtry.guide(npred, rf.type, spread.upper = 5, spread.lower = 4)
}
\arguments{
\item{npred}{Numeric. Input number of predictor variables.}

\item{rf.type}{One of: \code{"class"} for classification trees or \code{"reg"} for regression trees.}

\item{spread.upper}{Upper range of mtry to use above optimal mtry value, Default: 5}

\item{spread.lower}{Lower range of mtry to use below optimal mtry value: 4}
}
\value{
Returns a list: \cr
 \code{ideal.mtry} is optimal mtry based on tree type and number of predictor variables \cr
 \code{upper.mtry} is upper range of mtry for downstream hyperparameter tuning \cr
 \code{lower.mtry} is lowest value of mtry for downstream hyperparameter tuning \cr
 \code{range} is range of mtry values to try in downstream hyperparameter tuning given in the format: \code{upper.mtry:lower.mtry}
}
\description{
Recommended mtry range based on number of paramerets and tree type
}
\details{
Written primarily for rf wrapper function
}
