% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_Functions.R
\name{split.ratio}
\alias{split.ratio}
\title{Split dataset}
\usage{
\method{split}{ratio}(df, nsets = 10, train.ratio = 0.8, rf.type, vpred)
}
\arguments{
\item{df}{Input dataset as dataframe.}

\item{nsets}{Number of random sets to generate, Default: 10}

\item{train.ratio}{Proportion of sample assigned to training set. Range from 0 to 1. E.g 0.8 indicates 80% of sample assigned to training set for a 80:20 train:test split, Default: 0.8}

\item{rf.type}{One of: \code{"class"} for classification trees or \code{"reg"} for regression trees.}

\item{vpred}{Name of response variable. Required if rf.type = class.}
}
\value{
Outputs nsets*2 number of dataframes. Training sets are labeled train.i and Hold-out/Test sets are labeled hold.i where i ranges from 1 to nsets.
}
\description{
Randomly split dataset by some ratio (e.g 80 train:20 test). Repeat several time for n number of sets. Default is 80:20.
}
\details{
Written primarily for rf wrapper function.
}
